@@ -12,55 +12,63 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.0",
-    "@angular/cdk": "^5.2.0",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/material": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
+    "@angular-devkit/build-webpack": "^0.13.8",
+    "@angular/animations": "7.2.14",
+    "@angular/cdk": "^7.3.7",
+    "@angular/common": "7.2.14",
+    "@angular/compiler": "7.2.14",
+    "@angular/core": "7.2.14",
+    "@angular/fire": "^5.1.2",
+    "@angular/forms": "7.2.14",
+    "@angular/http": "7.2.14",
+    "@angular/material": "^7.3.7",
+    "@angular/platform-browser": "7.2.14",
+    "@angular/platform-browser-dynamic": "7.2.14",
+    "@angular/router": "7.2.14",
     "@mdi/angular-material": "^3.4.93",
     "@mdi/font": "^3.6.95",
     "@ngx-translate/core": "^9.1.1",
     "@ngx-translate/http-loader": "^2.0.1",
     "angular2-oauth2": "^1.3.10",
-    "angularfire2": "^4.0.0-rc.0",
-    "angularx-qrcode": "^1.5.3",
+    "angularfire2": "^5.1.2",
+    "angularx-qrcode": "^1.6.0",
     "core-js": "^2.4.1",
     "file-saver": "^2.0.0",
-    "firebase": "^3.9.*",
+    "firebase": "^5.10.0",
     "json-server": "^0.14.0",
     "lodash": "^4.17.11",
     "moment": "^2.24.0",
     "ng2-pdf-viewer": "^5.2.3",
     "ng4-loading-spinner": "^1.2.3",
     "pbkdf2": "^3.0.17",
-    "rxjs": "^5.5.6",
+    "rxjs": "^6.5.1",
     "rxjs-compat": "^6.3.3",
-    "zone.js": "^0.8.19"
+    "tslib": "^1.9.0",
+    "webpack-dev-server": "^3.3.1",
+    "zone.js": "^0.8.29"
   },
   "devDependencies": {
-    "@angular/cli": "~1.7.4",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
-    "@types/jasmine": "~2.8.3",
+    "@angular-devkit/build-angular": "^0.13.8",
+    "@angular/cli": "^7.3.8",
+    "@angular/compiler-cli": "7.2.14",
+    "@angular/language-service": "7.2.14",
+    "@types/jasmine": "^3.3.12",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
-    "jasmine-core": "~2.8.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
-    "karma-chrome-launcher": "~2.2.0",
+    "@types/node": "^11.13.5",
+    "codelyzer": "^5.0.0",
+    "jasmine-core": "^3.4.0",
+    "jasmine-spec-reporter": "^4.2.1",
+    "karma": "^4.1.0",
+    "karma-chrome-launcher": "^2.2.0",
+    "karma-cli": "^2.0.0",
     "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "^5.4.1",
+    "karma-jasmine": "^2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "protractor": "^5.4.2",
+    "rxjs-tslint": "^0.1.7",
     "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "tslint": "^5.16.0",
+    "typescript": "3.2.4",
+    "webpack": "^4.30.0"
   }
 }
