@@ -3,9 +3,8 @@ import { FormGroup, FormBuilder, Validators } from '@angular/forms';
 import { Router, ActivatedRoute } from '@angular/router';
 import { AuthService } from '../../../account/services/auth.service';
 import { TranslateService } from '@ngx-translate/core';
-import { Subscription } from 'rxjs/Subscription';
+import { Subscription ,  BehaviorSubject } from 'rxjs';
 import { AccountLocalConstant } from '../account-locale.constant';
-import { BehaviorSubject } from 'rxjs';
 import { Ng4LoadingSpinnerService } from 'ng4-loading-spinner';
 import { NotificationService } from '../../../shared/services/notification.service';
 import { Location } from '@angular/common';
