@@ -1,7 +1,7 @@
 export class Inventory {
   public static tabs = [
-    { text: 'inventory.material.title', icon: 'view_list', link: '/inventory' },
+    // { text: 'inventory.material.title', icon: 'view_list', link: '/inventory' },
     { text: 'inventory.product.title', icon: 'assignment', link: '/product-list' },
-    { text: 'inventory.kitchen.title', icon: 'silverware', link: '/kitchen' }
+    // { text: 'inventory.kitchen.title', icon: 'silverware', link: '/kitchen' }
   ];
 }
